Qvain light fix (#217)
* Fixed error related to Marshmallow version.

* Removed uiId from participant data that is sent to backend, and moved parts to its own file.
